backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.40.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
scikit-learn==1.6.0
numpy==2.2.0
python-multipart==0.0.18
pydantic==2.10.3
pydantic-settings==2.7.0
slowapi==0.1.9
click==8.1.8
rich==13.9.4
